Module core_ethereum_types::constants

Constants

INDEXER_BLOCK_RANGE

Default initial block range used to query the RPC provider e.g. starting to query logs of 2000 blocks, if that fails, try with 1000 blocks etc.

INDEXER_TIMEOUT

Time the indexer waits to confirm a transaction

MAX_TRANSACTION_BACKOFF

Submitting a transaction get retried using an exponential backoff The last try should not take longer than MAX_TRANSACTION_BACKOFF

PROVIDER_CACHE_TTL

specifies in milliseconds for how long various ethereum requests like eth_getBalance should be cached for

TX_CONFIRMATION_WAIT

Time to wait for a confirmation before giving up If the gas price is too low, the indexer would otherwise wait forever.